Load sample data: blue whale lunge-feeding

These sample data are provided to gain familiarity with Stickleback functionality. They contain bio-logging sensor data and labeled behavioral events for six feeding blue whales.

load_lunges()

Value

[list(Sensors, Events)] A list with two named elements. sensorsis a Sensors object with bio-logging sensor data collected from feeding blue whales and events is an Events object with the times of feeding events.

Details

Blue whales (Balaenoptera musculus) prey on krill schools using a behavior called "lunge feeding". Lunges have a stereotypical kinematic signature that makes them readily identifiable in bio-logging sensor data. First, the animal accelerates from about 2 to 4-5 m/s, then it opens its mouth and decelerates to almost a complete stop (<1 m/s). Blue whales usually lunge from below, so the pitch is often positive (~30-45 deg) and sometimes they roll on their side (~90 deg) or even do a barrel roll (0-360 deg). See Cade et al. (2016), Goldbogen et al. (2006), and Goldbogen et al. (2017) for more details.

The bio-logging sensor data made available by load_lunges contains 1.5 - 2 hour records from six blue whales tagged on September 4th and 5th, 2018. The 11.5 hours of data include 218 lunge feeding events, and were previously published by Goldbogen et al. (2019).

The examples in sb_plot_data, sb_fit, sb_predict, and sb_assess show how to use these sample data.

References

Cade, D. E., Friedlaender, A. S., Calambokidis, J., & Goldbogen, J. A. (2016). Kinematic diversity in rorqual whale feeding mechanisms. Current Biology, 26(19), 2617-2624.

Goldbogen, J. A., Calambokidis, J., Shadwick, R. E., Oleson, E. M., McDonald, M. A., & Hildebrand, J. A. (2006). Kinematics of foraging dives and lunge-feeding in fin whales. Journal of Experimental Biology, 209(7), 1231-1244.

Goldbogen, J. A., Cade, D. E., Calambokidis, J., Friedlaender, A. S., Potvin, J., Segre, P. S., & Werth, A. J. (2017). How baleen whales feed: the biomechanics of engulfment and filtration. Annual Review of Marine Science, 9, 367-386.

Goldbogen, J. A., Cade, D. E., Wisniewska, D. M., Potvin, J., Segre, P. S., Savoca, M. S., ... & Pyenson, N. D. (2019). Why whales are big but not bigger: Physiological drivers and ecological limits in the age of ocean giants. Science, 366(6471), 1367-1372.